requirement.txt
fastapi
torch
transformers
huggingface-hub
python-dotenv